style: chance form style

+++ src/Entity/Service.php
@@ -8,7 +8,7 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 
-#[ORM\Entity(repositoryClass: ServiceRepository::class)]
+#[ORM\Entity( repositoryClass: ServiceRepository::class )]
 class Service
 {
     #[ORM\Id]
@@ -16,40 +16,40 @@ class Service
     #[ORM\Column]
     private ?int $id = null;
 
-    #[ORM\Column(length: 255)]
+    #[ORM\Column( length: 255 )]
     private ?string $name = null;
 
-    #[ORM\Column(type: Types::TEXT, nullable: true)]
+    #[ORM\Column( type: Types::TEXT, nullable: true )]
     private ?string $description = null;
 
-    #[ORM\ManyToMany(targetEntity: CategoryService::class, inversedBy: 'services')]
+    #[ORM\ManyToMany( targetEntity: CategoryService::class, inversedBy: 'services' )]
     private Collection $categories;
 
-    #[ORM\Column(nullable: true)]
-    private ?int $price = null;
+    #[ORM\Column( type: "decimal", precision: 10, scale: 2, nullable: true )]
+    private ?float $price = null;
 
-    #[ORM\Column(type: Types::TIME_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::TIME_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $duration = null;
 
-    #[ORM\Column(type: Types::TIME_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::TIME_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $startTime = null;
 
-    #[ORM\Column(type: Types::TIME_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::TIME_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $endTime = null;
 
-    #[ORM\Column(length: 255, nullable: true)]
+    #[ORM\Column( length: 255, nullable: true )]
     private ?string $status = null;
 
-    #[ORM\Column(type: Types::DATE_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::DATE_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $startDate = null;
 
-    #[ORM\Column(type: Types::DATE_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::DATE_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $endDate = null;
 
-    #[ORM\Column(nullable: true)]
+    #[ORM\Column( nullable: true )]
     private ?int $avalaibleSpacePerService = null;
 
-    #[ORM\Column(type: Types::TIME_MUTABLE, nullable: true)]
+    #[ORM\Column( type: Types::TIME_MUTABLE, nullable: true )]
     private ?\DateTimeInterface $bufferTime = null;
 
     public function __construct()
@@ -57,29 +57,29 @@ class Service
         $this->categories = new ArrayCollection();
     }
 
-    public function getId(): ?int
+    public function getId() : ?int
     {
         return $this->id;
     }
 
-    public function getName(): ?string
+    public function getName() : ?string
     {
         return $this->name;
     }
 
-    public function setName(string $name): self
+    public function setName( string $name ) : self
     {
         $this->name = $name;
 
         return $this;
     }
 
-    public function getDescription(): ?string
+    public function getDescription() : ?string
     {
         return $this->description;
     }
 
-    public function setDescription(?string $description): self
+    public function setDescription( ?string $description ) : self
     {
         $this->description = $description;
 
@@ -89,129 +89,129 @@ class Service
     /**
      * @return Collection<int, CategoryService>
      */
-    public function getCategories(): Collection
+    public function getCategories() : Collection
     {
         return $this->categories;
     }
 
-    public function addCategory(CategoryService $category): self
+    public function addCategory( CategoryService $category ) : self
     {
-        if (!$this->categories->contains($category)) {
-            $this->categories->add($category);
+        if ( !$this->categories->contains( $category ) ) {
+            $this->categories->add( $category );
         }
 
         return $this;
     }
 
-    public function removeCategory(CategoryService $category): self
+    public function removeCategory( CategoryService $category ) : self
     {
-        $this->categories->removeElement($category);
+        $this->categories->removeElement( $category );
 
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice() : ?float
     {
         return $this->price;
     }
 
-    public function setPrice(?int $price): self
+    public function setPrice( ?float $price ) : self
     {
         $this->price = $price;
 
         return $this;
     }
 
-    public function getDuration(): ?\DateTimeInterface
+    public function getDuration() : ?\DateTimeInterface
     {
         return $this->duration;
     }
 
-    public function setDuration(?\DateTimeInterface $duration): self
+    public function setDuration( ?\DateTimeInterface $duration ) : self
     {
         $this->duration = $duration;
 
         return $this;
     }
 
-    public function getStartTime(): ?\DateTimeInterface
+    public function getStartTime() : ?\DateTimeInterface
     {
         return $this->startTime;
     }
 
-    public function setStartTime(?\DateTimeInterface $startTime): self
+    public function setStartTime( ?\DateTimeInterface $startTime ) : self
     {
         $this->startTime = $startTime;
 
         return $this;
     }
 
-    public function getEndTime(): ?\DateTimeInterface
+    public function getEndTime() : ?\DateTimeInterface
     {
         return $this->endTime;
     }
 
-    public function setEndTime(?\DateTimeInterface $endTime): self
+    public function setEndTime( ?\DateTimeInterface $endTime ) : self
     {
         $this->endTime = $endTime;
 
         return $this;
     }
 
-    public function getStatus(): ?string
+    public function getStatus() : ?string
     {
         return $this->status;
     }
 
-    public function setStatus(?string $status): self
+    public function setStatus( ?string $status ) : self
     {
         $this->status = $status;
 
         return $this;
     }
 
-    public function getStartDate(): ?\DateTimeInterface
+    public function getStartDate() : ?\DateTimeInterface
     {
         return $this->startDate;
     }
 
-    public function setStartDate(?\DateTimeInterface $startDate): self
+    public function setStartDate( ?\DateTimeInterface $startDate ) : self
     {
         $this->startDate = $startDate;
 
         return $this;
     }
 
-    public function getEndDate(): ?\DateTimeInterface
+    public function getEndDate() : ?\DateTimeInterface
     {
         return $this->endDate;
     }
 
-    public function setEndDate(?\DateTimeInterface $endDate): self
+    public function setEndDate( ?\DateTimeInterface $endDate ) : self
     {
         $this->endDate = $endDate;
 
         return $this;
     }
 
-    public function getAvalaibleSpacePerService(): ?int
+    public function getAvalaibleSpacePerService() : ?int
     {
         return $this->avalaibleSpacePerService;
     }
 
-    public function setAvalaibleSpacePerService(?int $avalaibleSpacePerService): self
+    public function setAvalaibleSpacePerService( ?int $avalaibleSpacePerService ) : self
     {
         $this->avalaibleSpacePerService = $avalaibleSpacePerService;
 
         return $this;
     }
 
-    public function getBufferTime(): ?\DateTimeInterface
+    public function getBufferTime() : ?\DateTimeInterface
     {
         return $this->bufferTime;
     }
 
-    public function setBufferTime(?\DateTimeInterface $bufferTime): self
+    public function setBufferTime( ?\DateTimeInterface $bufferTime ) : self
     {
         $this->bufferTime = $bufferTime;
